// src/types/profile.ts
export interface UserProfile {
    profile_image: string;
    full_name: string;
    email: string;
    phone: string;
    address: string;
    zip_code: string;
    hobby: string;
    bio: string;
    maiden_name: string;
    date_of_birth: string;
    gender: string;
    nationality: string;
    country: string;
  }
  